Factor NodePool into its own package
